assistants_api_mcp/requirements.txt
openai>=1.35.0
python-dotenv>=1.0.1
mcp>=0.1.0
